Doug comments
